+++ src/community/infomap.cpp
@@ -18,20 +18,19 @@
 
 #include "igraph_community.h"
 
-#include "igraph_error.h"
+#include "config.h"
+
+#ifdef HAVE_INFOMAP
+
 #include "igraph_interface.h"
 #include "igraph_interrupt.h"
 
 #include "core/exceptions.h"
 
-#include "config.h"
-
-#ifdef HAVE_INFOMAP
-    #include "Infomap.h"
-#endif
-
 #include <climits>
 #include <cmath>
+
+#include "Infomap.h"
 
 static igraph_error_t infomap_get_membership(infomap::InfomapBase &infomap, igraph_vector_int_t *membership) {
     igraph_int_t n = infomap.numLeafNodes();
@@ -110,6 +109,7 @@ static igraph_error_t convert_igraph_to_infomap(const igraph_t *graph,
 static bool infomap_allow_interruption() {
     return igraph_allow_interruption();
 }
+#endif // HAVE_INFOMAP
 
 /**
  * \function igraph_community_infomap